# docs/swagger.yaml
basePath: /api
definitions:
  credit-report-service_internal_apperr.ErrorBody:
    properties:
      details: {}
      error:
        type: string
      message:
        type: string
      status:
        type: integer
      timestamp:
        type: string
    type: object
  credit-report-service_internal_models.Account:
    properties:
      createdAt:
        type: string
      dateOfBirth:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      phone:
        type: string
      profileCompleted:
        type: boolean
      role:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  credit-report-service_internal_models.CreditAnalyticsRequest:
    properties:
      accountId:
        type: integer
      clientRefNum:
        type: string
      createdAt:
        type: string
      httpStatus:
        type: integer
      id:
        type: integer
      message:
        type: string
      mobileNo:
        type: string
      requestBody:
        items:
          type: integer
        type: array
      requestId:
        type: string
      responseBody:
        items:
          type: integer
        type: array
      resultCode:
        type: integer
    type: object
...
  credit-report-service_internal_models.KYCRecord:
    properties:
      aadhaarLast4:
        type: string
      aadhaarPanLinked:
        type: boolean
      aadhaarReference:
        type: string
      accountId:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      pan:
        type: string
      panName:
        type: string
      panVerified:
        type: boolean
      provider:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      verifiedAt:
        type: string
    type: object
  credit-report-service_internal_models.Order:
    properties:
      amount:
        type: number
      cfOrderId:
        type: string
      createdAt:
        type: string
      currency:
        type: string
      failureReason:
        type: string
      orderId:
        type: string
      paidAt:
        type: string
      paymentMethod:
        type: string
      paymentSessionId:
        type: string
      productCode:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
  credit-report-service_internal_models.Product:
    properties:
      active:
        type: boolean
      amount:
        type: number
      code:
        type: string
      createdAt:
        type: string
      currency:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  credit-report-service_internal_service.AuthResult:
    properties:
      account:
        $ref: '#/definitions/credit-report-service_internal_models.Account'
      expiresAt:
        type: string
      refreshToken:
        type: string
      token:
        type: string
    type: object
  credit-report-service_internal_service.CreditAnalyticsInput:
    properties:
      device_ip:
        type: string
    type: object
  credit-report-service_internal_service.PurchaseResult:
    properties:
      amount:
        type: number
      cfOrderId:
        type: string
      currency:
        type: string
      mode:
        type: string
      orderId:
        type: string
      paymentSessionId:
        type: string
      status:
        type: string
    type: object
  credit-report-service_internal_service.ReportPage:
    properties:
      items:
        items:
          $ref: '#/definitions/credit-report-service_internal_service.ReportSummary'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
  credit-report-service_internal_service.SessionView:
    properties:
      current:
        description: |-
          Current marks the session the requesting device is using right now, so
          the UI can label it and warn before revoking it.
        type: boolean
      deviceInfo:
        allOf:
        - $ref: '#/definitions/credit-report-service_internal_models.DeviceMeta'
        description: |-
          DeviceInfo is the make/model/OS breakdown, split into what the client
          declared and what the server parsed from the User-Agent.
      deviceName:
        type: string
      expiresAt:
        type: string
      id:
        type: integer
      ip:
        type: string
      lastUsedAt:
        type: string
      platform:
        type: string
      signedInAt:
        type: string
    type: object
  credit-report-service_internal_service.SignupResult:
    properties:
      accountId:
        type: integer
      email:
        type: string
      message:
        type: string
    type: object
  internal_handler.createOrderReq:
    properties:
      productCode:
        example: CREDIT_ANALYSIS
        type: string
    type: object
  internal_handler.googleLoginReq:
    properties:
      idToken:
        example: eyJhbGciOiJSUzI1NiIs...
        type: string
    type: object
  internal_handler.loginReq:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: hunter2pass
        type: string
    type: object
  internal_handler.refreshReq:
    properties:
      refreshToken:
        example: rt_8Kx3...
        type: string
    type: object
  internal_handler.resendReq:
    properties:
      email:
        example: user@example.com
        type: string
    type: object
  internal_handler.signupReq:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: hunter2pass
        type: string
    type: object
  internal_handler.submitPanReq:
    properties:
      pan:
        example: ABCDE1234F
        type: string
    type: object
...
  internal_handler.verifyEmailReq:
    properties:
      email:
        example: user@example.com
        type: string
      otp:
        example: "123456"
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API for the credit-report Android app: email/password auth with OTP verification, JWT sessions, profile management, and credit-report CRUD.

    All routes are mounted under `/api`. Protected routes require an `Authorization: Bearer <jwt>` header obtained from login or email verification.
  title: Credit Report Service API
  version: "1.0"
paths:
...
  /auth/google:
    post:
      consumes:
      - application/json
      description: Verifies a Google ID token (issued by the Android/iOS Google Sign-In
        SDK) and returns a session JWT. On first login, creates a verified account;
        on subsequent logins, reuses the existing account. If the Google email matches
        an existing account, the Google identity is linked onto it. Requires the Web
        OAuth client ID to be configured (AUTH_GOOGLE_CLIENT_ID); otherwise returns
        503.
      parameters:
      - description: Stable per-device UUID
        in: header
        name: X-Device-Id
        type: string
      - description: Human-readable device name shown in the device list
        in: header
        name: X-Device-Name
        type: string
      - description: ios | android | web
        in: header
        name: X-Device-Platform
        type: string
      - description: JSON device description, e.g. {\
        in: header
        name: X-Device-Info
        type: string
      - description: Google ID token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.googleLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit-report-service_internal_service.AuthResult'
        "400":
          description: Validation failed (missing idToken)
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "401":
          description: Invalid Google token
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "409":
          description: Google identity already linked to another account
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "503":
          description: Google login is not configured
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      summary: Log in with Google
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: 'Verifies email + password and opens a session for the calling
        device. Returns a short-lived access JWT plus a refresh token — in the JSON
        body for mobile clients, or as an httpOnly `refresh_token` cookie when `X-Device-Platform:
        web`. Send `X-Device-Id` (a stable per-device UUID) so repeat logins update
        one entry in the device list instead of creating a new one.'
      parameters:
      - description: Stable per-device UUID
        in: header
        name: X-Device-Id
        type: string
      - description: Human-readable device name shown in the device list
        in: header
        name: X-Device-Name
        type: string
      - description: ios | android | web
        in: header
        name: X-Device-Platform
        type: string
      - description: JSON device description, e.g. {\
        in: header
        name: X-Device-Info
        type: string
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit-report-service_internal_service.AuthResult'
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "401":
          description: Invalid email or password / email not verified
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      summary: Log in with email + password
      tags:
      - auth
  /auth/logout:
    post:
      description: Revokes the session the access token was issued for and clears
        the web refresh cookie. The access token itself stays valid until it expires
        (minutes), so clients should discard it too.
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\": \"Signed out\"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      security:
      - BearerAuth: []
      summary: Sign out of the current device
      tags:
      - auth
  /auth/otp/resend:
    post:
      consumes:
      - application/json
      description: Re-issues the signup OTP for an unverified email, subject to cooldown
        / send limits.
      parameters:
      - description: Email to resend to
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.resendReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"message\": \"Verification code re-sent\"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation failed / cooldown / send limit
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "404":
          description: No signup found for this email
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "409":
          description: Email is already verified
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      summary: Resend signup verification OTP
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 'Rotates the refresh token and issues a fresh access token. Mobile
        clients send the token in the JSON body; web clients send nothing — the httpOnly
        `refresh_token` cookie is read automatically. The old refresh token is invalidated
        on every call, so replaying one revokes the whole session as a theft signal.
        Public endpoint: the refresh token itself is the credential.'
      parameters:
      - description: Stable per-device UUID
        in: header
        name: X-Device-Id
        type: string
      - description: ios | android | web
        in: header
        name: X-Device-Platform
        type: string
      - description: JSON device description, e.g. {\
        in: header
        name: X-Device-Info
        type: string
      - description: Refresh token (omit for web cookie clients)
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_handler.refreshReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit-report-service_internal_service.AuthResult'
        "401":
          description: Missing, expired, replayed, or unknown refresh token
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      summary: Exchange a refresh token for a new token pair
      tags:
      - auth
  /auth/sessions:
    delete:
      description: Revokes all of the account's sessions except the caller's own.
        Use after a password change or when a device is lost.
      produces:
      - application/json
      responses:
        "200":
          description: '{\"revoked\": 3}'
          schema:
            additionalProperties:
              type: integer
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      security:
      - BearerAuth: []
      summary: Sign out every other device
      tags:
      - auth
    get:
      description: 'Returns the account''s active sessions, most recently used first.
        The entry matching the caller''s own access token is flagged with `current:
        true`.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/credit-report-service_internal_service.SessionView'
            type: array
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      security:
      - BearerAuth: []
      summary: List signed-in devices
      tags:
      - auth
...
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a PENDING account with an unverified password identity
        and emails a verification OTP. Re-signing up for an unverified email updates
        the password and re-issues the OTP.
      parameters:
      - description: Signup credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.signupReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/credit-report-service_internal_service.SignupResult'
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "409":
          description: Email already registered
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      summary: Sign up with email + password
      tags:
      - auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Checks the signup OTP; on success verifies the identity, activates
        the account, and opens a session for the calling device. Token delivery follows
        the same rules as POST /auth/login.
      parameters:
      - description: Stable per-device UUID
        in: header
        name: X-Device-Id
        type: string
      - description: Human-readable device name shown in the device list
        in: header
        name: X-Device-Name
        type: string
      - description: ios | android | web
        in: header
        name: X-Device-Platform
        type: string
      - description: JSON device description, e.g. {\
        in: header
        name: X-Device-Info
        type: string
      - description: Email + OTP
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.verifyEmailReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit-report-service_internal_service.AuthResult'
        "400":
          description: Wrong / expired / locked OTP
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "409":
          description: Email already registered
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      summary: Verify email with OTP
      tags:
      - auth
  /credit-analytics/reports:
    get:
      description: Returns a paginated list of the caller's reports, newest first.
        Each item carries only the report id (unique identifier) and generation date.
      parameters:
      - default: 1
        description: 1-indexed page number (default 1)
        in: query
        name: page
        type: integer
      - default: 20
        description: page size (default 20, max 100)
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit-report-service_internal_service.ReportPage'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      security:
      - BearerAuth: []
      summary: List the authenticated account's credit-analytics reports
      tags:
      - credit-analytics
...
  /credit-analytics/request:
    post:
      consumes:
      - application/json
      description: Builds the Digitap /credit_analytics/request payload from the authenticated
        account's profile and KYC record (mobile, name, PAN), plus server-generated
        values (client_ref_num, otp, timestamp), then persists the request and full
        upstream response and returns the stored row. Only device_ip is taken from
        the request body; if omitted it falls back to the detected remote IP.
      parameters:
      - description: Credit-analytics request (device_ip optional; defaults to the
          caller's IP)
        in: body
        name: request
        schema:
          $ref: '#/definitions/credit-report-service_internal_service.CreditAnalyticsInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/credit-report-service_internal_models.CreditAnalyticsRequest'
        "400":
          description: Invalid request body / missing profile or PAN / upstream 400
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "401":
          description: Not authenticated / upstream 401
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "422":
          description: Upstream tradeline limit exceeded
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      security:
      - BearerAuth: []
      summary: Request a credit analysis from Digitap
      tags:
      - credit-analytics
  /kyc/pan:
    post:
      consumes:
      - application/json
      description: Accepts a PAN (Permanent Account Number), validates the format,
        and upserts it against the account's KYC record. A re-submission overwrites
        any existing PAN and resets verification to PENDING.
      parameters:
      - description: PAN to store
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.submitPanReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/credit-report-service_internal_models.KYCRecord'
        "400":
          description: Invalid JSON body / PAN format
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "409":
          description: PAN already linked to another account
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      security:
      - BearerAuth: []
      summary: Submit the authenticated account's PAN
      tags:
      - kyc
  /orders:
    get:
      description: Returns every order placed by the authenticated account, newest
        first.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/credit-report-service_internal_models.Order'
            type: array
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      security:
      - BearerAuth: []
      summary: List the authenticated account's orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Creates an order for the given product against the authenticated
        account and registers it with Cashfree. The response carries `paymentSessionId`
        and `mode`, which the frontend passes to the Cashfree JS SDK to open checkout.
      parameters:
      - description: Product to purchase
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createOrderReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/credit-report-service_internal_service.PurchaseResult'
        "400":
          description: Invalid JSON body / productCode missing or unknown
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "409":
          description: Product is not available for purchase
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "502":
          description: Payment gateway could not create the order
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create an order and start checkout
      tags:
      - orders
  /orders/{orderId}:
    get:
      description: Returns a single order owned by the authenticated account. If the
        order is still awaiting payment its status is reconciled against Cashfree
        before being returned, so this is the endpoint to poll after checkout closes.
      parameters:
      - description: Order id returned from order creation
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit-report-service_internal_models.Order'
        "400":
          description: orderId is not valid
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get one order
      tags:
      - orders
  /payments/cashfree/webhook:
    post:
      consumes:
      - application/json
      description: 'Server-to-server endpoint called by Cashfree on payment success/failure.
        Not for client use: there is no bearer auth, trust comes from the `x-webhook-signature`
        HMAC over the raw body, and `x-idempotency-key` de-duplicates redeliveries.'
      parameters:
      - description: Cashfree signature timestamp
        in: header
        name: x-webhook-timestamp
        required: true
        type: string
      - description: Base64 HMAC-SHA256 of timestamp+body
        in: header
        name: x-webhook-signature
        required: true
        type: string
      - description: Cashfree delivery id, used to drop duplicates
        in: header
        name: x-idempotency-key
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid webhook payload
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "401":
          description: Invalid webhook signature
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      summary: Cashfree payment webhook
      tags:
      - payments
...
  /products:
    get:
      description: Returns the active product catalog with prices. Use a product's
        `code` as the `productCode` when creating an order.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/credit-report-service_internal_models.Product'
            type: array
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      security:
      - BearerAuth: []
      summary: List purchasable products
      tags:
      - orders
  /profile:
    get:
      description: Returns the full account record for the current session.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit-report-service_internal_models.Account'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get the authenticated account's profile
      tags:
      - profile
    put:
      consumes:
      - application/json
      description: Sets first/last name (required) and optional date of birth (YYYY-MM-DD),
        marking the profile step complete.
      parameters:
      - description: Profile fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credit-report-service_internal_models.Account'
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/credit-report-service_internal_apperr.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update the authenticated account's profile
      tags:
      - profile
securityDefinitions:
  BearerAuth:
    description: '"Bearer <jwt>" — obtain via POST /api/auth/login or /api/auth/verify-email'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"